[build] dto들 수정

+++ src/main/java/Goweb/FormMaker/dto/survey/CreateResponseDto.java
@@ -11,6 +11,6 @@ public class CreateResponseDto {
 
     private Long questionId;
     private String answer;
-    private Set<Integer> responseOptions; // optionId
+    private Set<Long> responseOptions; // optionId
 
 }
